// src/common/enums/document.enum.ts
export enum DocumentOwnerType {
  TRUCK = 'truck',
  TRAILER = 'trailer',
  DRIVER = 'driver',
  COMPANY = 'company',
}

export enum DocumentCategory {
  INSURANCE = 'insurance',
  VTV = 'vtv',
  LICENSE = 'license',
  ID_CARD = 'id_card',
  PERMIT = 'permit',
  DELIVERY_NOTE = 'delivery_note',
  WAYBILL = 'waybill',
  OTHER = 'other',
}

export enum DocumentStatus {
  VALID = 'valid',
  EXPIRING = 'expiring',
  EXPIRED = 'expired',
}